# Conflicts:
#	gk-war/src/main/webapp/static/dist/before/scripts/teacher-lecture.min.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/gk_hot.min.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/gk_hot_detile.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/gk_hot_detile.min.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/profession_info.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/profession_info.min.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/school_info.js
#	gk-war/src/main/webapp/static/dist/consult/scripts/school_info.min.js
#	gk-war/src/main/webapp/static/dist/consult/styles/school_info.min.css
#	gk-war/src/main/webapp/static/dist/index/scripts/index.js
#	gk-war/src/main/webapp/static/dist/index/scripts/index.min.js
#	gk-war/src/main/webapp/static/dist/user/scripts/personal-info.js
#	gk-war/src/main/webapp/static/dist/user/scripts/personal-info.min.js
#	gk-war/src/main/webapp/static/dist/user/styles/expert-service-detail.min.css
#	gk-war/src/main/webapp/static/dist/user/styles/expert-service.min.css
#	gk-war/src/main/webapp/static/dist/user/styles/order-expert.min.css
#	gk-war/src/main/webapp/static/dist/user/styles/personal-info.min.css

+++ consult-service/src/main/java/cn/thinkjoy/gk/service/impl/MajoredServiceImpl.java
@@ -21,7 +21,7 @@ public class MajoredServiceImpl implements IMajoredService {
     private IMajoredDAO majoredDAO;
 
     @Override
-    public List<Map<String, Object>> getMajoreByParentId(int i) {
+    public List<Map<String, Object>> getMajoreByParentId(Long i) {
         return majoredDAO.getMajoreByParentId(i);
     }
